requirements.txt
scikit-learn==1.6.1
xgboost==2.1.4
lightgbm==4.6.0
numpy==1.26.4
scipy==1.14.1
pytorch_tabnet
tabpfn

lightautoml

catboost==1.2.7
lightgbm==4.6.0
cloudpickle
optuna
scikit-lego

autogluon.tabular
ray
dask